feat: loading skeleton on games
