--- app/games/loading.tsx
@@ -1,3 +1,0 @@
-export default function GamesLoading() {
-  return <>Loading...</>
-}
