+++ app/games/styles.module.css
@@ -11,24 +11,7 @@
   flex-direction: column;
   width: 100%;
 }
-.list {
-  margin: 0;
-  padding: 0rem;
-  list-style: none;
-  width: 75%;
-  display: flex;
-  flex-direction: column;
-}
-.list li:nth-of-type(2n + 1) {
-  width: 100%;
-  display: flex;
-  justify-content: flex-start;
-}
-.list li:nth-of-type(2n) {
-  width: 100%;
-  display: flex;
-  justify-content: flex-end;
-}
+
 .title {
   color: var(--Primary-Color);
   text-align: center;
@@ -57,38 +40,4 @@
 .button:hover {
   background: var(--Secondary-Color);
   color: var(--Accent-Color);
-}
-@media (max-width: 720px) {
-  .list {
-    width: 100%;
-  }
-  .list li:nth-of-type(2n + 1) {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-  }
-  .list li:nth-of-type(2n) {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-  }
-}
-
-@media (min-width: 1150px) {
-  .list {
-    display: grid;
-    grid-template-columns: 50% 50%;
-    width: 100%;
-  }
-  .list li:nth-of-type(2n + 1),
-  .list li:nth-of-type(2n) {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    grid-row: span 2;
-  }
-  .list li:nth-child(2) {
-    grid-row: 2 / span 2;
-    grid-column: 2;
-  }
 }
