+++ components/Game/styles.module.css
@@ -27,8 +27,6 @@
   flex-basis: 100%;
   box-shadow: 3px 7px 5px #0001;
   border: 1px solid #3331;
-  transform-origin: center;
-  animation: fadeIn 0.3s ease-out forwards;
   background: var(--Secondary-Color);
 }
 .container:hover .editIcon {
@@ -43,6 +41,7 @@
   margin: 5px;
   border-radius: 5px;
   overflow: hidden;
+  background-color: var(--Placeholder-color);
   box-shadow: 3px 3px 5px rgba(0, 0, 0, 0.1);
   border: 1px solid #3331;
 }
@@ -65,15 +64,42 @@
   margin-right: 2rem;
 }
 
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-    transform: scale(0.5);
-  }
-  to {
-    opacity: 1;
-    transform: scale(1);
-  }
+.data_skeleton {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  padding: 1rem 0;
+  min-height: 140px;
+}
+
+.title_skeleton {
+  position: relative;
+  overflow: hidden;
+  width: 85%;
+  height: 1.5rem;
+  background-color: var(--Placeholder-color);
+}
+
+.paragraph_skeleton {
+  position: relative;
+  overflow: hidden;
+  width: 40%;
+  height: 1rem;
+  background-color: var(--Placeholder-color);
+}
+
+.loading_bar {
+  position: absolute;
+  height: 100%;
+  width: 50px;
+  background: linear-gradient(
+    to right,
+    var(--Placeholder-color),
+    white 65%,
+    white 85%,
+    var(--Placeholder-color)
+  );
+  animation: loading 1s linear infinite;
 }
 
 @media (max-width: 450px) {
@@ -93,5 +119,14 @@
   }
   .steam {
     margin: 0.5rem 0;
+  }
+}
+
+@keyframes loading {
+  from {
+    transform: translate(-35px);
+  }
+  to {
+    transform: translate(300px);
   }
 }
